+++ app/Http/Controllers/Admin/TagsController.php
@@ -3,13 +3,14 @@
 namespace App\Http\Controllers\Admin;
 
 use App\Http\Controllers\Controller;
+use App\Models\Admin\Product;
 use App\Models\Admin\Tag;
 use Illuminate\Http\Request;
 use Inertia\Inertia;
 
 class TagsController extends Controller
 {
-    
+
 
     public function index(Request $request)
     {
@@ -23,17 +24,16 @@ class TagsController extends Controller
     public function store(Request $request)
     {
         $request->validate([
-            'name' => 'required|string|max:255',     
+            'name' => 'required|string|max:255',
         ]);
 
-try{
-        Tag::create([
-            'name' => $request->name,
-        ]);
+        try {
+            Tag::create([
+                'name' => $request->name,
+            ]);
 
-        return back()->with('success', 'تم اضافة تصنيف جديد بنجاح');
-}
-        catch (\Exception $e) {
+            return back()->with('success', 'تم اضافة تصنيف جديد بنجاح');
+        } catch (\Exception $e) {
             return back()->with('error', 'حدث خطأ أثناء إضافة التصنيف: ' . $e->getMessage());
         }
     }
@@ -44,31 +44,40 @@ try{
             'name' => 'required|string|max:255',
         ]);
 
-       
-try{
 
-        $tag->update([
-            'name' => $request->name,
-        ]);
+        try {
 
-        return back()->with('success', 'تم تحديث التصنيف بنجاح');
-}
-        catch (\Exception $e) {
+            $tag->update([
+                'name' => $request->name,
+            ]);
+
+            return back()->with('success', 'تم تحديث التصنيف بنجاح');
+        } catch (\Exception $e) {
             return back()->with('error', 'حدث خطأ أثناء تحديث التصنيف: ' . $e->getMessage());
         }
     }
 
     public function destroy(Tag $tag)
     {
-try{
-        
+        try {
+            $hasProducts = Product::whereJsonContains(
+                'tags',
+                $tag->id
+            )->exists();
 
-        $tag->delete();
+            if ($hasProducts) {
 
-        return back()->with('success', 'تم حذف التصنيف بنجاح');
-}
-        catch (\Exception $e) {
-            return back()->with('error', 'حدث خطأ أثناء حذف التصنيف: ' . $e->getMessage());
+                return back()->with(
+                    'error',
+                    'لا يمكن حذف التصنيف لأن هناك لوحات مرتبطة به'
+                );
+            }
+
+            $tag->delete();
+
+            return back()->with('success', 'تم حذف التصنيف بنجاح');
+        } catch (\Exception $e) {
+            return back()->with('error', 'حدث خطأ أثناء حذف التصنيف');
         }
     }
 }
